Group manager buttons are missing  (#12426)
http://www.webgui.org/community/webgui-8/issues/12426

Ticket indicates otherwise, but the buttons appeared but didn't work, throwing
a JS error.

Yikes.	That thing is scary.  Code is in
www/extras/yui-webgui/build/form/groupManager.js.  It was written in 2011 by
Colin and touched only once, soon after that.  I haven't tested all of its
functionality because I'm afraid of what I'll find.  I cannot fathom how it
could do what it purports to, of adding multiple groups to a place where one
group is requested.  Does it go create a new group on the fly and dump all of
them in there?

Anyway, it looks like this got busted when the root/import/adminconsole got
moved to its own iframe.  groupManager.js looks for a tag with
id=application_workarea and fails to find it in that iframe, and it appears to
be part of root/import/adminconsole.  This means that this control only
previously worked while admin mode was on.  Group selection controls used in
other edit screens wouldn't have this functionality, at least not in a working
state.

Worked around the problem by resurrecting some dead code that looked for
the first head tag and changed it to look for the first body tag, and
then defaulted to that if the tag with id=application_workarea isn't
found.  The dialog pops up now, and does some stuff without blowing up.